docker/llm-service/requirements.txt
# LLM Service dependencies
fastapi
uvicorn[standard]
pydantic
python-dotenv
langchain
langchain-google-genai
httpx
python-multipart
